santa_cruz_obs_data/requirements.txt
numpy
netCDF4
thredds_crawler>=1.2.0
requests
epic2cf>=0.0.4
pyaxiom>=1.3.1
pytz
pynco
pytest
coloredlogs
compliance-checker
